requirements.txt
feedparser==6.0.11
pydantic==2.7.4
python-dotenv==1.0.1
